+++ Controller/booksController.js
@@ -4,13 +4,13 @@ const { Book, sequelize } = require('../models/');
 
 const booksController = {
     index: async (request, response) => {
-        let books =  await Book.findAll();
-        
+        let books = await Book.findAll();
+
         return response.json(books);
     },
 
     create: async (request, response) => {
-        let {name, author, description, publisher, generes_id, users_id} = request.body;
+        let { name, author, description, publisher, generes_id, users_id } = request.body;
 
         let newBook = await Book.create({
             name,
@@ -18,16 +18,14 @@ const booksController = {
             description,
             publisher,
             generes_id,
-            users_id,
-            // generes_id1, // pode estar errado no banco & é necessario na hora do create ou é apenas uma foreingKey
-            //users_id1:  // pode estar errado no banco & é necessario na hora do create ou é apenas uma foreingKey
+            users_id
         });
 
         return response.json(newBook);
     },
     update: async (request, response) => {
         let { id } = request.params;
-        let {name, author, description, publisher, generes_id, users_id} = request.body;
+        let { name, author, description, publisher, generes_id, users_id } = request.body;
 
         let updatedBook = await Book.update({
             name,
@@ -43,18 +41,18 @@ const booksController = {
         return response.send(updatedBook);
     },
     delete: async (request, response) => {
-       
+
         let { id } = request.params;
 
         const deletedBook = await Book.destroy({
-            where: {id}
+            where: { id }
         });
 
         return response.json(deletedBook);
-        
+
     },
 
-    show: async(request, response) => {
+    show: async (request, response) => {
         const { id } = request.params;
 
         const booksforUser = await Book.findAll({
