+++ models/Genere.js
@@ -11,7 +11,7 @@ module.exports = (sequelize, DataTypes) =>{
 
     Genere.associate = (models) => {
         Genere.hasMany(models.Book, { as: "book", foreignKey: "generes_id" });
-        Genere.belongsTo(models.Interest, { as: "interest", foreignKey: "generes_id" });
+        Genere.hasMany(models.Interest, { as: "interest", foreignKey: "generes_id" });
     };
 
     return Genere;
