+++ models/Interest.js
@@ -5,13 +5,13 @@ module.exports = (sequelize, DataTypes) =>{
             users_id: DataTypes.INTEGER,
             generes_id: DataTypes.INTEGER
         }, {
-            tableName: "interest",
+            tableName: "interests",
             timestamps: false
         }
     );
 
     Interest.associate = (models) => {
-        Interest.hasMany(models.Genere, { as: "genere", foreignKey: "generes_id" });
+        Interest.belongsTo(models.Genere, { as: "genere", foreignKey: "generes_id" });
         Interest.belongsTo(models.User, { as: "user", foreignKey: "users_id" });
     };
