+++ models/User.js
@@ -7,7 +7,8 @@ module.exports = (sequelize, DataTypes) =>{
             email: DataTypes.STRING,
             date_of_birth: DataTypes.STRING,
             gender: DataTypes.STRING,
-            phone_number: DataTypes.STRING
+            phone_number: DataTypes.STRING,
+            address_id: DataTypes.INTEGER
         }, {
             tableName: "users",
             timestamps: false
@@ -16,7 +17,7 @@ module.exports = (sequelize, DataTypes) =>{
 
     User.associate = (models) => {
         User.hasOne(models.Address, { as: "address", foreignKey: "users_id" });
-        User.hasMany(models.Interest, { as: "interest", foreignKey: "users_id" });
+        User.hasMany(models.Interest, { as: "interests", foreignKey: "users_id" });
         User.hasMany(models.Book, { as: "book", foreignKey: "users_id" });
     };
